+++ x-pack/platform/plugins/shared/lens/public/shared_components/legend/legend_settings.tsx
